Asgardio connected, next - connecting  IDP to frontend

+++ Vehicle-Service/src/main/java/com/vehicleservicereservator/VehicleService/ReservationService/ReservationServiceImpl.java
@@ -102,11 +102,16 @@ public class ReservationServiceImpl implements ReservationService{
 
         VehicleService vehicleService= vehicleServiceRepo.getReservationByID(book_id);
         DateValidator dateValidator = new DateValidator();
-        if (dateValidator.isDateTodayOrFuture(vehicleService.getDate(),vehicleService.getTime())) {
-             vehicleServiceRepo.deleteReservation(book_id);
-            return "Reservation Deleted Successfully" ;
+        if (vehicleService!=null){
+            if (dateValidator.isDateTodayOrFuture(vehicleService.getDate(),vehicleService.getTime())) {
+                vehicleServiceRepo.deleteReservation(book_id);
+                return "Reservation Deleted Successfully" ;
+            }else {
+                return "Cannot delete past reservations";
+            }
         }else {
-            return "Cannot delete past reservations";
+            return "Ha Ha Please Enter available reservation";
         }
+
     }
 }
